refactor(priority_queue): split into package (config, queue, processor) with re-exports
Modularize priority queue into `src/meridian/core/priority_queue/` and remove monolithic module. Preserve API used by scheduler; tests pass.
