+++ src/meridian/core/priority_queue/__init__.py
@@ -1,0 +1,11 @@
+from __future__ import annotations
+
+from .config import PriorityQueueConfig
+from .queue import PrioritySchedulingQueue
+from .processor import NodeProcessor
+
+__all__ = [
+    "PriorityQueueConfig",
+    "PrioritySchedulingQueue",
+    "NodeProcessor",
+]
